fix: link name was unable to find the thruster link
The thruster link couldn't be found since the previous
iteration of the code tried to find the value by
brute forcing searches until a link was found however
it also did the first iteration from the full name
twice on the first run which resulted on it being
unable to find the thruster model on one of our sdfs.
To avoid this behaviour, now the getLinkFromName searches
for a single link name, removing the bruteforce behaviour.