chore: implement element re-scoping based on a plugin name that does not suck
It turns out that Gazebo does accept plugin names with `::` and the name, and
that the built-in rescoping mechanism in ruby_sdformat already does re-scope as
it should. Removing the override of the plugin name in rock-gazebo, we end up
with a plugin name that is scoped relatively to the model root.

This commit refactors the utility API around this issue, and implements using
that (new-actually-old) way to name plugins to our advantage